Revert rename class to classes.
