+++ apps/vor/pkg/class/class.go
@@ -1,4 +1,4 @@
-package classes
+package class
 
 import (
 	"net/http"
